Move Debug Intrinsics
The debug intrinsics are intel specific and should be located in the
intel specific folder with the proper naming.
