+++ bfintrinsics/tests/arch/intel_x64/test_drs.cpp
@@ -19,7 +19,7 @@
 #include <catch/catch.hpp>
 #include <intrinsics.h>
 
-using namespace x64;
+using namespace intel_x64;
 
 dr7::value_type g_dr7 = 0;
